Restructuring slidepuzzle/main.js
This is just reordering the methods in the file, renaming some methods and adding/modifying some comments. Part of this was replacing "completed" with "solved" throughout, the code and comments were inconsistent on the use of that term. This also impacted the HTML, the div element now is named "puzzleSolved".