Updating main.js, playlist.json to reflect only seasons 1-5
On 6/5 at 3:00 PM I saw that S2E6 "Opie's Hobo Friend" was playing. Doing some calculations with the recent baseline updates it seems that Pluto shortened the sequence to 79.5 hours, 159 episodes. That is the number of episodes in seasons 1-5.

Changes to main.js only use 159 episodes in calculations. Also change to playlist.json for the same along with the new baseline from this airing.