Updating playlist.json to reflect new start time
After recent changes to the airings on Pluto I noticed on 6/8 that the start times were five minutes earlier, not "on the hour/half hour". I monitored it for a while and they do not seem to be moving back, so I am updating the baseline time in playlist.json to 06/08/2022 2:55am (central).