Making build-and-deploy run on pushes to master
Before making this change changed how GitHub Pages were deployed for this repo (under the "Settings" tab, on the left is a "Pages" section). The setting is under "Build and deployment" and was labeled "Source", changed it to "Use GitHub Actions" from "Deploy from a branch" (was set to master, of course).

https://github.blog/changelog/2022-07-27-github-pages-custom-github-actions-workflows-beta/