Modifying slidepuzzle to allow any tile to be the initial empty tile
The initial empty tile (currently always the last) is saved when the puzzle is randomized & restored when the puzzle is solved.

The puzzle tiles are now numbered sequentially 0..8, empty tile is 9.

Updating isPuzzleSolved() and checkAndMarkPuzzleSolved() to reflect that the empty cell may not always be the last.

Also making makeNewPuzzle() a little more 'generic'
- Changing how the tileNums array is created, no longer a hard-coded literal array
- Initial tile number no longer hard-coded literal
- Renaming some locals
- Added more comments