Fixing display issues on Chrome mobile for Pluto site
With the switch to w3-css the text was very small in Portrait orientation, found the fix was to remove "-webkit-text-size-adjust" that w3-css adds to the <html> element style.

Also changing how the button is displayed, it was always too small on Chrome mobile. For some reason Chrome mobile displays <button> elements with very small text. Switching it to an anchor element with styles to make it look like a w3-button.