Fixing initial display of Update button for Pluto site
On Chrome mobile the text (and height) of the Update button would be smaller than other text in the body when initially loaded in portrait orientation. If you switched to landscape then back, then the text would be the same size as other text in the body.

I have theories on why Chrome is doing this, but regardless, setting the text of the button after displaying the listing "fixes" this.

Also changing the style of the update button, getting rid of the "hover" (because on mobile it stays in the hover color after pressing) and adding "ripple" (which changes the color only while pressing). Also adding inline style to not change the mouse cursor.

Also changing how error messages are displayed from main.js. A new function formats the HTML as a list item so that it looks better.