Switching to display Pluto listings as cards
Switching the listing element back to being a plain-old <div>, the entries inside are styled as w3-card-2 with a blue heading. Removing the w3-ccc-grey styles that are no longer used.

Note: While debugging I found that the setting of the Update button text at the end of updateListing() to "fix" the formatting is no longer needed (presumably due to the way the html is now structured). There is some inconsistency in Chrome so I added a comment to the code and left it in.