Updating playlist.json for changing sequence of Season 5
Noticed on 11/11/2022 that S4E15 was playing, playlist said early S1 should be playing
(I had checked early AM 11/06/22 that the the playlist was correct, this was the day of the time change)

After monitoring through the end of S4 I saw that S5E1 played next. This is the first time I had seen this episode, they used to show S5E6 as the first episode of S5.