Updating playlist.json for changing sequence of Season 4
Noticed on 12/04/2022 that S4E22 "Andy's Vacation" was playing, when the playlist said E21 "The Shoplifters" should be re-airing. After that S5E23 "Andy Saves Gomer" was shown, this was the first time I had seen this on pluto.
(I did not catch this part of S4 when I was monitoring the S5 changes on 11/11, possibly this change was made at the same time)