Updating playlist.json for new sequence start time
I noticed on 12/11 that the episode that was playing didn't match my playlist, so I decided to keep monitoring it. I watched for a while and noticed on 12/19 that it changed again. I monitored the airings through (roughly) two cycles to see that it didn't change again.
Calculated baseline: 12/19/2022 6:30:00 AM -06:00